:truck: Move utils into utils
