+++ docs/functions/weather/utils.ts
@@ -1,0 +1,128 @@
+export const supportedLocales = ["en", "de"] as const;
+
+type WWOCode =
+  | 0
+  | 1
+  | 2
+  | 3
+  | 45
+  | 48
+  | 51
+  | 53
+  | 55
+  | 56
+  | 57
+  | 61
+  | 63
+  | 65
+  | 66
+  | 67
+  | 71
+  | 73
+  | 75
+  | 77
+  | 80
+  | 81
+  | 82
+  | 85
+  | 86
+  | 95
+  | 96
+  | 99;
+
+export const wmoCodeToEmojiMap: Record<WWOCode, string> = {
+  0: "☀️", // Clear sky
+  1: "🌤️", // Mainly clear
+  2: "⛅", // Partly cloudy
+  3: "☁️", // Overcast
+  45: "🌫️", // Fog
+  48: "🌫️", // Depositing rime fog
+  51: "🌧️", // Drizzle: Light intensity
+  53: "🌧️", // Drizzle: Moderate intensity
+  55: "🌧️", // Drizzle: Dense intensity
+  56: "🌧️", // Freezing Drizzle: Light intensity
+  57: "🌧️", // Freezing Drizzle: Dense intensity
+  61: "🌧️", // Rain: Slight intensity
+  63: "🌧️", // Rain: Moderate intensity
+  65: "🌧️", // Rain: Heavy intensity
+  66: "🌧️", // Freezing Rain: Light intensity
+  67: "🌧️", // Freezing Rain: Heavy intensity
+  71: "❄️", // Snow fall: Slight intensity
+  73: "❄️", // Snow fall: Moderate intensity
+  75: "❄️", // Snow fall: Heavy intensity
+  77: "❄️", // Snow grains
+  80: "🌧️", // Rain showers: Slight intensity
+  81: "🌧️", // Rain showers: Moderate intensity
+  82: "🌧️", // Rain showers: Violent intensity
+  85: "❄️", // Snow showers: Slight intensity
+  86: "❄️", // Snow showers: Heavy intensity
+  95: "⛈️", // Thunderstorm: Slight or moderate
+  96: "⛈️", // Thunderstorm with slight hail
+  99: "⛈️", // Thunderstorm with heavy hail
+};
+
+type WMOTranslation = Record<WWOCode, string>;
+
+export const wmoCodeToText: Record<(typeof supportedLocales)[number], WMOTranslation> =
+  {
+    en: {
+      0: "Clear sky",
+      1: "Mainly clear",
+      2: "Partly cloudy",
+      3: "Overcast",
+      45: "Fog",
+      48: "Depositing rime fog",
+      51: "Light drizzle",
+      53: "Moderate drizzle",
+      55: "Dense drizzle",
+      56: "Light freezing drizzle",
+      57: "Dense freezing drizzle",
+      61: "Slight rain",
+      63: "Moderate rain",
+      65: "Heavy rain",
+      66: "Light freezing rain",
+      67: "Heavy freezing rain",
+      71: "Slight snow fall",
+      73: "Moderate snow fall",
+      75: "Heavy snow fall",
+      77: "Snow grains",
+      80: "Slight rain showers",
+      81: "Moderate rain showers",
+      82: "Violent rain showers",
+      85: "Slight snow showers",
+      86: "Heavy snow showers",
+      95: "Slight or moderate thunderstorm",
+      96: "Thunderstorm with slight hail",
+      99: "Thunderstorm with heavy hail",
+    },
+    de: {
+      0: "Klarer Himmel",
+      1: "Überwiegend klar",
+      2: "Teilweise bewölkt",
+      3: "Bedeckt",
+      45: "Nebel",
+      48: "Rauhreifnebel",
+      51: "Leichter Nieselregen",
+      53: "Mäßiger Nieselregen",
+      55: "Dichter Nieselregen",
+      56: "Leichter gefrierender Nieselregen",
+      57: "Dichter gefrierender Nieselregen",
+      61: "Leichter Regen",
+      63: "Mäßiger Regen",
+      65: "Starker Regen",
+      66: "Leichter gefrierender Regen",
+      67: "Starker gefrierender Regen",
+      71: "Leichter Schneefall",
+      73: "Mäßiger Schneefall",
+      75: "Starker Schneefall",
+      77: "Schneekörner",
+      80: "Leichte Regenschauer",
+      81: "Mäßige Regenschauer",
+      82: "Heftige Regenschauer",
+      85: "Leichte Schneeschauer",
+      86: "Starke Schneeschauer",
+      95: "Leichtes oder mäßiges Gewitter",
+      96: "Gewitter mit leichtem Hagel",
+      99: "Gewitter mit starkem Hagel",
+    },
+  };
